backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
python-dateutil
gunicorn
dj-database-url
